+++ crates/mtr-packet/src/backend/mod.rs
@@ -12,9 +12,7 @@ use crate::probe_table::ProbeTable;
 
 #[cfg(test)]
 pub mod fake;
-#[cfg(target_os = "macos")]
-pub mod macos;
-#[cfg(any(target_os = "linux", target_os = "freebsd"))]
+#[cfg(any(target_os = "linux", target_os = "freebsd", target_os = "macos"))]
 pub mod unix;
 
 pub trait ProbeBackend {